Add ProjectsComponent
Added the ProjectsComponent to the website. Added additional routes to
the router to include the new Projects route and changed the catchall to
redirect to the home page.

Fixed the routerLinkActive of the Home link to only be active when the
route is an exact match with the home page route.